moving the changes made
